// src/app/features/auth/models/cliente.model.ts
export class ClienteModel {
  idCliente: number = 0;
  identificacion: number = 0;
  nombre: string = "";
  direccion: string = "";
  idDepartamento: number = 0;
  idMunicipio: number = 0;
  telefono: number = 0;
  correoElectronico: string = '';
  contrasena: string = "";
  idAfiliado: number = 0;
  idBanco: number = 0;
  cuentaBancaria: number = 0;
  codigoAfiliado: string = "";
  codigoReferido: string = ""
  idAfiliadoPadre: number = 0;
}